+++ entity_parser_key_parser_test.go
@@ -739,6 +739,7 @@ func TestIndexParse(t *testing.T) {
 		InputIndexName    string
 		ExpectedIndexName string
 		PrimaryKey        *PrimaryKey
+		Columns           []string
 		Error             error
 	}{
 		{
@@ -925,16 +926,50 @@ func TestIndexParse(t *testing.T) {
 			InputIndexName: "searchByKey",
 			Error:          errors.New("index name (searchByKey) must be exported, try (SearchByKey) instead"),
 		},
+		{
+			Tag:               "name=jj key=ok columns=(ok)",
+			ExpectedIndexName: "jj",
+			PrimaryKey: &PrimaryKey{
+				PartitionKeys:  []string{"ok"},
+				ClusteringKeys: nil,
+			},
+			InputIndexName: "SearchByKey",
+			Columns:        []string{"ok"},
+			Error:          nil,
+		},
+		{
+			Tag:               "name=jj key=ok columns=(ok, test, hi,)",
+			ExpectedIndexName: "jj",
+			PrimaryKey: &PrimaryKey{
+				PartitionKeys:  []string{"ok"},
+				ClusteringKeys: nil,
+			},
+			InputIndexName: "SearchByKey",
+			Columns:        []string{"ok", "test", "hi"},
+			Error:          nil,
+		},
+		{
+			Tag:               "name=jj key=ok columns=(ok, test, (hi),)",
+			ExpectedIndexName: "jj",
+			PrimaryKey: &PrimaryKey{
+				PartitionKeys:  []string{"ok"},
+				ClusteringKeys: nil,
+			},
+			InputIndexName: "SearchByKey",
+			Columns:        []string{"ok", "test", "hi"},
+			Error:          errors.New("index field SearchByKey with an invalid dosa index tag: columns=(ok, test, (hi),)"),
+		},
 	}
 
 	for _, d := range data {
-		name, primaryKey, err := parseIndexTag(d.InputIndexName, d.Tag)
+		name, primaryKey, columns, err := parseIndexTag(d.InputIndexName, d.Tag)
 		if d.Error != nil {
 			assert.Contains(t, err.Error(), d.Error.Error())
 		} else {
 			assert.Nil(t, err)
 			assert.Equal(t, name, d.ExpectedIndexName)
 			assert.Equal(t, primaryKey, d.PrimaryKey)
+			assert.Equal(t, columns, d.Columns)
 		}
 	}
 }